pyproject.toml
[project]
name = "ph-civic-data-mcp"
version = "0.3.0"
description = "MCP server for Philippine civic data: PSGC, infra spending accountability, earthquakes, weather, typhoons, procurement, population, poverty, solar/climate, air quality, vegetation"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Xavier Puspus" }]
keywords = ["mcp", "philippines", "phivolcs", "pagasa", "philgeps", "psa", "psgc", "nasa", "modis", "usgs", "ibtracs", "world-bank", "civic-tech", "accountability"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
    "fastmcp>=3.0.0,<4.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.9.0",
    "cachetools>=5.3.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.urls]
Homepage = "https://github.com/xmpuspus/ph-civic-data-mcp"
Repository = "https://github.com/xmpuspus/ph-civic-data-mcp"
Issues = "https://github.com/xmpuspus/ph-civic-data-mcp/issues"

[project.scripts]
ph-civic-data-mcp = "ph_civic_data_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ph_civic_data_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
markers = [
    "live: hits real upstream endpoints; opt-in via -m live",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint.per-file-ignores]
"docs/live_probe_v020.py" = ["E402"]
"docs/live_probe_v030.py" = ["E402"]
"docs/live_demo.py" = ["E402"]
"docs/live_demo_single.py" = ["E402"]
"docs/live_demo_v030.py" = ["E402"]
